Refactor scrape.
